feat: add versions page component
Move versions page logic to dedicated component for better organization
