+++ src/routes/versions.tsx
@@ -1,65 +1,9 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { FolderPlusIcon } from "lucide-react";
-import { AnimatePresence } from "motion/react";
 
-import { MotionVersionContextMenu } from "@/components/context-menus/version.context-menu";
-import { AddVersionDialog } from "@/components/dialogs/addversion.dialog";
-import { VersionRow } from "@/components/rows/version.row";
-import { Button } from "@/components/ui/button";
-import { DialogTrigger } from "@/components/ui/dialog";
+import { SettingsPage } from "@/components/pages/settings";
 import { ErrorComponent } from "@/components/ui/error";
-import { ScrollArea } from "@/components/ui/scroll-area";
-import { useInstalledVersions } from "@/hooks/use-installed-versions";
-import { compareSemverDesc, itemVariants } from "@/lib/utils";
-
-import { rootDialogHandle } from "./__root";
 
 export const Route = createFileRoute("/versions")({
-  component: RouteComponent,
+  component: SettingsPage,
   errorComponent: ErrorComponent,
 });
-
-export function RouteComponent() {
-  const { data: versions } = useInstalledVersions();
-
-  const sorted = (versions ?? []).toSorted((a, b) => compareSemverDesc(a.name, b.name));
-
-  return (
-    <div className="grid w-full grid-rows-[min-content_auto] gap-2">
-      <div className="flex h-fit gap-2 pt-1 pr-2 pl-2 max-md:pl-9">
-        <Button
-          className="w-full cursor-pointer justify-between"
-          render={<DialogTrigger handle={rootDialogHandle} payload={() => <AddVersionDialog />} />}
-          variant="outline"
-        >
-          <span className="flex text-xs">Add version</span>
-          <FolderPlusIcon className="size-4" />
-        </Button>
-      </div>
-      <ScrollArea className="h-full px-2">
-        <AnimatePresence>
-          {sorted.map((version, index) => (
-            <MotionVersionContextMenu
-              animate="show"
-              className="flex items-center gap-2 p-2 not-last:border-b"
-              custom={index}
-              exit="exit"
-              initial="hidden"
-              key={`${version.name}-context-menu`}
-              layout="position"
-              variants={itemVariants}
-              version={version.name}
-            >
-              <VersionRow version={version} />
-            </MotionVersionContextMenu>
-          ))}
-          {sorted.length === 0 && (
-            <p className="text-muted-foreground p-4 text-sm select-none">
-              No versions yet. Click "Add version" to get started.
-            </p>
-          )}
-        </AnimatePresence>
-      </ScrollArea>
-    </div>
-  );
-}
